Route : modification du compte

+++ back/controllers/comments.js
@@ -27,21 +27,21 @@ const globalFunctions = require('../global/functions');
  */
 // Publication d'un commentaire.
 exports.newComment = (req, res) => {
-    const newMessageValidator = new Validator(req.body, {
+    const newCommentValidator = new Validator(req.body, {
         linkedMessage: 'required|integer|maxLength:11',
         content: 'required|string|maxLength:3000'
     });
-    globalFunctions.areVariablesValid(newMessageValidator).then(areVariablesValid => {
+    globalFunctions.areVariablesValid(newCommentValidator).then(areVariablesValid => {
         if (areVariablesValid === false) {
             return res.status(400).json({ error: globalVariables.ERROR_WRONG_DATA });
         }
         const userId = req.headers.user_id,         /* Variable déjà vérifiée par le middleware 'auth.js' */
         { content, linkedMessage } = req.body;
-        globalFunctions.findOneMessage(linkedMessage).then((message) => {
+        globalFunctions.findOneMessage(linkedMessage).then(message => {
             if (message === null) {
                 return res.status(400).json({ error: globalVariables.ERROR_WRONG_DATA });
             }
-            Comment.findOne({ where: { userId: userId }, order: [[ 'id', 'DESC' ]] }).then((comment) => {
+            Comment.findOne({ where: { userId: userId }, order: [[ 'id', 'DESC' ]] }).then(comment => {
                 if (comment !== null && comment.timestamp >= globalVariables.CURRENT_TIMESTAMP - 60) {          /* On autorise un commentaire par minute */
                     return res.status(400).json({ error: globalVariables.ERROR_WRONG_DATA });
                 }
@@ -64,11 +64,11 @@ exports.getComments = (req, res) => {
             return res.status(400).json({ error: globalVariables.ERROR_WRONG_DATA });
         }
         const id = req.params.id;
-        globalFunctions.findOneMessage(id).then((message) => {
+        globalFunctions.findOneMessage(id).then(message => {
             if (message === null) {
                 return res.status(400).json({ error: globalVariables.ERROR_WRONG_DATA });
             }
-            Comment.findAll({ where: { linkedMessage: id }, order: [[ 'id', 'DESC' ]] }).then((comments) => {
+            Comment.findAll({ where: { linkedMessage: id }, order: [[ 'id', 'DESC' ]] }).then(comments => {
                 res.status(200).json(comments);
             });
         });
@@ -84,7 +84,7 @@ exports.editComment = (req, res) => {
         const userId = req.headers.user_id,         /* Variable déjà vérifiée par le middleware 'auth.js' */
         id = req.params.id,
         content = req.body.content;
-        Comment.update({ content: content }, { where: { id: id, userId: userId }, limit: 1 }).then((comment) => {
+        Comment.update({ content: content }, { where: { id: id, userId: userId }, limit: 1 }).then(comment => {
             if (comment === 0) {
                 return res.status(400).json({ error: globalVariables.ERROR_WRONG_DATA });
             }
@@ -103,14 +103,14 @@ exports.delComment = (req, res) => {
         id = req.params.id;
         globalFunctions.isAdmin(userId).then(isAdmin => {
             if (isAdmin === true) {
-                Comment.destroy({ where: { id: id }, limit: 1 }).then((comment) => {
+                Comment.destroy({ where: { id: id }, limit: 1 }).then(comment => {
                     if (comment === 0) {
                         return res.status(400).json({ error: globalVariables.ERROR_WRONG_DATA });
                     }
                     res.status(200).json({ message: globalVariables.SUCCESS });
                 });
             } else {
-                Comment.destroy({ where: { id: id, userId: userId }, limit: 1 }).then((comment) => {
+                Comment.destroy({ where: { id: id, userId: userId }, limit: 1 }).then(comment => {
                     if (comment === 0) {
                         return res.status(400).json({ error: globalVariables.ERROR_WRONG_DATA });
                     }
